fix(web): guard non-string command template to prevent TypeError crash
detectCommandFromText called cmd.template.trim() after a truthy-only guard
(if (!cmd.template) continue). The opencode API (and MCP/skill-sourced
commands) can return a non-string truthy template (object/number/array),
which passed the guard and threw 'TypeError: e.template.trim is not a
function' on Safari/WebKit, crashing the session render.

Better Stack (Kortix Frontend prod, app 2346967):
- pattern 4efb3dc3... (6), 9ca5739b... (4), 5de308bb... (1)
- 'TypeError: e.template.trim is not a function. (In e.template.trim(...),
   e.template.trim is undefined/not a function)'
- last seen 2026-07-12 ~06:05 UTC

Fix: skip any command whose template is not a string (typeof guard) at the
boundary, so command detection degrades gracefully instead of throwing.
Empty/whitespace templates are still skipped downstream by the existing
prefix.length < 20 check, preserving prior behavior.

Also extracts the pure detectCommandFromText into its own React-free module
(detect-command.ts) so it can be unit-tested, and applies the same guard to
the mobile mirror in apps/mobile SessionTurn.tsx.

Regression tests (detect-command.test.ts) prove non-string (object/number/
array/bool/null/undefined) templates no longer throw and that a valid
command alongside malformed ones still matches.
